+++ client.ts
@@ -18,19 +18,19 @@
 
   // Set up session and clear the session param on load
   const url = new URL(location.href);
-  let session = url.searchParams.get('giscus');
+  let session = url.searchParams.get('giscus') || '';
   const savedSession = localStorage.getItem(GISCUS_SESSION_KEY);
   url.searchParams.delete('giscus');
+  url.hash = '';
   const cleanedLocation = url.toString();
 
   if (session) {
     localStorage.setItem(GISCUS_SESSION_KEY, JSON.stringify(session));
     history.replaceState(undefined, document.title, cleanedLocation);
-  } else {
+  } else if (savedSession) {
     try {
-      session = JSON.parse(savedSession || '') || '';
+      session = JSON.parse(savedSession);
     } catch (e) {
-      session = '';
       localStorage.removeItem(GISCUS_SESSION_KEY);
       console.warn(`${formatError(e?.message)} Session has been cleared.`);
     }
@@ -93,30 +93,29 @@
   // Set up iframe element
   const iframeElement = document.createElement('iframe');
   const iframeAttributes = {
-    class: 'giscus-frame',
+    class: 'giscus-frame giscus-frame--loading',
     title: 'Comments',
     scrolling: 'no',
+    allow: 'clipboard-write',
     src,
     loading,
   };
   Object.entries(iframeAttributes).forEach(
     ([key, value]) => value && iframeElement.setAttribute(key, value),
   );
+  // Prevent white flash on load
+  iframeElement.style.opacity = '0';
+  iframeElement.addEventListener('load', () => {
+    iframeElement.style.removeProperty('opacity');
+    iframeElement.classList.remove('giscus-frame--loading');
+  });
 
-  // Create default style and prepend as <head>'s first child to make override possible.
-  const style = document.getElementById('giscus-css') || document.createElement('style');
+  // Link default style and prepend as <head>'s first child to make override possible.
+  const style =
+    (document.getElementById('giscus-css') as HTMLLinkElement) || document.createElement('link');
   style.id = 'giscus-css';
-  style.textContent = `
-  .giscus, .giscus-frame {
-    width: 100%;
-    min-height: 150px;
-  }
-
-  .giscus-frame {
-    border: none;
-    color-scheme: normal;
-  }
-`;
+  style.rel = 'stylesheet';
+  style.href = `${giscusOrigin}/default.css`;
   document.head.prepend(style);
 
   // Insert iframe element
@@ -132,6 +131,12 @@
   }
   const suggestion = `Please consider reporting this error at https://github.com/giscus/giscus/issues/new.`;
 
+  function signOut() {
+    delete params.session;
+    const src = `${giscusOrigin}/widget?${new URLSearchParams(params)}`;
+    iframeElement.src = src; // Force reload
+  }
+
   // Listen to messages
   window.addEventListener('message', (event) => {
     if (event.origin !== giscusOrigin) return;
@@ -143,19 +148,27 @@
       iframeElement.style.height = `${data.giscus.resizeHeight}px`;
     }
 
+    if (data.giscus.signOut) {
+      localStorage.removeItem(GISCUS_SESSION_KEY);
+      console.log(`[giscus] User has logged out. Session has been cleared.`);
+      signOut();
+      return;
+    }
+
     if (!data.giscus.error) return;
 
     const message: string = data.giscus.error;
 
-    if (message.includes('Bad credentials') || message.includes('Invalid state value')) {
+    if (
+      message.includes('Bad credentials') ||
+      message.includes('Invalid state value') ||
+      message.includes('State has expired')
+    ) {
       // Might be because token is expired or other causes
       if (localStorage.getItem(GISCUS_SESSION_KEY) !== null) {
         localStorage.removeItem(GISCUS_SESSION_KEY);
         console.warn(`${formatError(message)} Session has been cleared.`);
-
-        delete params.session;
-        const src = `${giscusOrigin}/widget?${new URLSearchParams(params)}`;
-        iframeElement.src = src; // Force reload
+        signOut();
       } else if (!savedSession) {
         console.error(`${formatError(message)} No session is stored initially. ${suggestion}`);
       }