+++ lib/variables.ts
@@ -7,6 +7,7 @@ export const env = {
   private_key: process.env.GITHUB_PRIVATE_KEY,
   encryption_password: process.env.ENCRYPTION_PASSWORD,
   app_host: process.env.NEXT_PUBLIC_GISCUS_APP_HOST as `https://${string}`,
+  revalidate_first_page: process.env.NEXT_PUBLIC_REVALIDATE_FIRST_PAGE !== 'false',
   origins: JSON.parse(process.env.ORIGINS || '[]') as string[],
   origins_regex: JSON.parse(process.env.ORIGINS_REGEX || '[]') as string[],
 } as const;
@@ -21,11 +22,12 @@ export const availableThemes = [
   'dark_protanopia',
   'dark_tritanopia',
   'dark_dimmed',
-  'transparent_dark',
   'preferred_color_scheme',
+  'transparent_dark',
+  'cobalt',
   'custom',
 ] as const;
 
-export type AvailableTheme = typeof availableThemes[number];
+export type AvailableTheme = (typeof availableThemes)[number];
 
 export type Theme = AvailableTheme | `/${string}` | `https://${string}`;